# apps/api/openapi.yaml
openapi: 3.1.0

info:
  title: Quiz App API
  version: 1.0.0
  description: >
    REST API for the Quiz App — create and manage quizzes, questions, and
    authentication.

    Resource IDs are UUIDs (strings). Timestamps are ISO 8601 UTC
    (e.g. 2026-04-25T10:00:00Z).

    Field names use camelCase throughout to match the frontend TypeScript
    types directly (no snake_case conversion required).

servers:
  - url: http://localhost:3000/api
    description: Local development

tags:
  - name: Auth
    description: Register, login, refresh, and logout.
  - name: Users
    description: Current user profile.
  - name: Quizzes
    description: Create, read, update, delete, publish, and duplicate quizzes.

paths:

  # ── Auth ─────────────────────────────────────────────────────────────────────

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: Plain-text password; hashed server-side before persistence.
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken, tokenType, expiresIn, user]
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  tokenType:
                    type: string
                    examples:
                      - Bearer
                  expiresIn:
                    type: integer
                    format: int32
                    description: Access token lifetime in seconds.
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/EmailConflict'

  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
            examples:
              default:
                value:
                  email: admin@quiz.app
                  password: admin1234
      responses:
        '200':
          description: Signed in successfully
          content:
            application/json:
              schema:
                type: object
                required: [message, user]
                properties:
                  message:
                    type: string
                    examples:
                      - Signed in successfully.
                  user:
                    type: object
                    required: [email]
                    properties:
                      email:
                        type: string
                        format: email
                        examples:
                          - admin@quiz.app
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken, tokenType, expiresIn, user]
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  tokenType:
                    type: string
                    examples:
                      - Bearer
                  expiresIn:
                    type: integer
                    format: int32
                    description: Access token lifetime in seconds.
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/oauth:
    post:
      tags: [Auth]
      summary: Sign in or register via OAuth (e.g. Google)
      description: >
        Finds an existing account by email or creates a new one if none exists.
        The created account has no usable password — it can only be accessed via OAuth.
        Returns a full JWT access + refresh token pair identical to the login response.
      operationId: oauthLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        '200':
          description: OAuth sign-in successful
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken, tokenType, expiresIn, user]
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  tokenType:
                    type: string
                    examples:
                      - Bearer
                  expiresIn:
                    type: integer
                    format: int32
                    description: Access token lifetime in seconds.
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate current refresh token
      operationId: logoutUser
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Users ─────────────────────────────────────────────────────────────────────

  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getMyProfile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me/avatar:
    put:
      tags: [Users]
      summary: Update current user avatar
      operationId: updateAvatar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  description: Base64-encoded image data (max ~200 KB).
      responses:
        '200':
          description: Avatar updated; returns updated user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Users]
      summary: Remove current user avatar
      operationId: removeAvatar
      responses:
        '200':
          description: Avatar removed; returns updated user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Quizzes ───────────────────────────────────────────────────────────────────

  /quizzes:
    get:
      tags: [Quizzes]
      summary: List all quizzes
      operationId: listQuizzes
      responses:
        '200':
          description: Array of all quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'

    post:
      tags: [Quizzes]
      summary: Create a quiz
      operationId: createQuiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, questions]
              properties:
                title:
                  type: string
                  maxLength: 120
                  examples:
                    - World Capitals Quiz
                description:
                  type: string
                  maxLength: 500
                  examples:
                    - Test your knowledge of world capitals.
                questions:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/QuestionInput'
      responses:
        '201':
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '400':
          $ref: '#/components/responses/ValidationError'

  /quizzes/{quizId}:
    parameters:
      - $ref: '#/components/parameters/QuizIdPath'

    get:
      tags: [Quizzes]
      summary: Get a quiz by ID
      operationId: getQuiz
      responses:
        '200':
          description: Quiz detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          $ref: '#/components/responses/QuizNotFound'

    patch:
      tags: [Quizzes]
      summary: Update a quiz
      operationId: updateQuiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: All fields are optional; only provided fields are updated.
              properties:
                title:
                  type: string
                  maxLength: 120
                  examples:
                    - Updated Quiz Title
                description:
                  type: string
                  maxLength: 500
                  examples:
                    - Updated description.
                questions:
                  type: array
                  items:
                    $ref: '#/components/schemas/QuestionInput'
      responses:
        '200':
          description: Quiz updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/QuizNotFound'

    delete:
      tags: [Quizzes]
      summary: Delete a quiz
      operationId: deleteQuiz
      responses:
        '204':
          description: Quiz deleted
        '404':
          $ref: '#/components/responses/QuizNotFound'

  /quizzes/{quizId}/publish:
    parameters:
      - $ref: '#/components/parameters/QuizIdPath'
    patch:
      tags: [Quizzes]
      summary: Publish a quiz
      description: Transitions the quiz status to `PUBLISHED`.
      operationId: publishQuiz
      responses:
        '200':
          description: Quiz published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          $ref: '#/components/responses/QuizNotFound'

  /quizzes/{quizId}/unpublish:
    parameters:
      - $ref: '#/components/parameters/QuizIdPath'
    patch:
      tags: [Quizzes]
      summary: Unpublish a quiz
      description: Transitions the quiz status to `UNPUBLISHED`.
      operationId: unpublishQuiz
      responses:
        '200':
          description: Quiz unpublished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          $ref: '#/components/responses/QuizNotFound'

  /quizzes/{quizId}/duplicate:
    parameters:
      - $ref: '#/components/parameters/QuizIdPath'
    post:
      tags: [Quizzes]
      summary: Duplicate a quiz
      description: >
        Creates a full copy of the quiz with a new ID.
        The copy's title is prefixed with "Copy of " and its status
        is reset to `in_progress`.
      operationId: duplicateQuiz
      responses:
        '201':
          description: Duplicate quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          $ref: '#/components/responses/QuizNotFound'

components:

  parameters:

    QuizIdPath:
      name: quizId
      in: path
      required: true
      description: UUID of the quiz.
      schema:
        type: string
      examples:
        default:
          value: 550e8400-e29b-41d4-a716-446655440000


  schemas:

    # ── Users ─────────────────────────────────────────────────────────────────

    UserProfile:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
        avatar:
          type: string
          nullable: true
          description: Base64-encoded avatar image, or null if not set.

    # ── Enums ─────────────────────────────────────────────────────────────────

    UserRole:
      type: string
      enum: [USER, ADMIN]

    QuizStatus:
      type: string
      enum:
        - IN_PROGRESS
        - PUBLISHED
        - UNPUBLISHED
      description: >
        `IN_PROGRESS` — default after creation;
        `PUBLISHED` — visible/active;
        `UNPUBLISHED` — hidden/inactive.

    # ── Domain objects ────────────────────────────────────────────────────────

    Question:
      type: object
      required: [prompt, options, correctOptionIndex]
      properties:
        id:
          type: string
          description: UUID assigned by the server.
          readOnly: true
          examples:
            - 3f2504e0-4f89-11d3-9a0c-0305e82c3301
        prompt:
          type: string
          maxLength: 500
          examples:
            - What is the capital of France?
        options:
          type: array
          minItems: 2
          items:
            type: string
          examples:
            - [Paris, London, Berlin, Madrid]
        correctOptionIndex:
          type: integer
          minimum: 0
          description: Zero-based index of the correct answer in `options`.
          examples:
            - 0

    Quiz:
      type: object
      required: [id, title, description, status, questions, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: UUID assigned by the server.
          readOnly: true
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          maxLength: 120
          examples:
            - World Capitals Quiz
        description:
          type: string
          maxLength: 500
          examples:
            - Test your knowledge of world capitals.
        status:
          $ref: '#/components/schemas/QuizStatus'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        createdAt:
          type: string
          format: date-time
          readOnly: true
          examples:
            - '2026-04-25T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          examples:
            - '2026-04-25T12:00:00.000Z'

    QuestionInput:
      type: object
      required: [prompt, options, correctOptionIndex]
      description: Question payload used in create/update requests.
      properties:
        prompt:
          type: string
          maxLength: 500
          examples:
            - What is the capital of France?
        options:
          type: array
          minItems: 2
          items:
            type: string
          examples:
            - [Paris, London, Berlin, Madrid]
        correctOptionIndex:
          type: integer
          minimum: 0
          description: Zero-based index of the correct answer in `options`.
          examples:
            - 0

    # ── Errors ────────────────────────────────────────────────────────────────

    ProblemDetail:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples:
            - VALIDATION_ERROR
        detail:
          type: string
          description: Human-readable explanation of this occurrence.
          examples:
            - Request validation failed.
        errors:
          type: array
          description: Field-level errors. Present on 400 responses only.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                examples:
                  - title
              message:
                type: string
                examples:
                  - must be shorter than or equal to 120 characters

  responses:

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                detail: Request validation failed.
                errors:
                  - field: title
                    message: must be shorter than or equal to 120 characters
                  - field: questions
                    message: must contain at least 1 elements

    Unauthorized:
      description: Invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: UNAUTHORIZED
                detail: Invalid email or password.

    EmailConflict:
      description: Email is already registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: EMAIL_ALREADY_EXISTS
                detail: The email address is already registered.
                errors:
                  - field: email
                    message: already registered

    QuizNotFound:
      description: Quiz not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: QUIZ_NOT_FOUND
                detail: No quiz found with the given ID.